requirements.txt
async-generator==1.10
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
nest-asyncio==1.5.6
outcome==1.2.0
psycopg==3.1.8
psycopg-binary==3.1.8
pynput==1.7.6
pyobjc-core==9.0.1
pyobjc-framework-ApplicationServices==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Quartz==9.0.1
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.2
selenium==4.8.3
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
sortedcontainers==2.4.0
splinter==0.19.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
websocket-client==1.5.1
websockets==11.0.1
wsproto==1.2.0
